Some jazz

+++ feature/overview/src/main/kotlin/uk/co/weightwars/overview/OverviewNavGraph.kt
@@ -3,7 +3,6 @@ package uk.co.weightwars.overview
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
-import androidx.navigation.toRoute
 import kotlinx.serialization.Serializable
 import uk.co.weightwars.overview.activeChallenge.ActiveChallengeScreen
 import uk.co.weightwars.overview.overviewScreen.OverviewScreen
@@ -12,7 +11,7 @@ import uk.co.weightwars.overview.overviewScreen.OverviewScreen
 object OverviewRoute
 
 @Serializable
-data class ActiveChallengeRoute(val activeChallengeId: Int)
+data class ActiveChallengeRoute(val activeChallengeId: Long)
 
 fun NavGraphBuilder.overviewNavGraph(navController: NavHostController) {
     composable<OverviewRoute> {
